Revert "Trends API"
