+++ complaint_search/es_builders.py
@@ -11,33 +11,6 @@ from complaint_search.defaults import (
 )
 
 
-def build_search_terms(search_term, field):
-    if (re.match(r"^[A-Za-z\d\s]+$", search_term) and not
-        any(keyword in search_term
-            for keyword in ("AND", "OR", "NOT", "TO"))):
-
-        # Match Query
-        return {
-            "match": {
-                field: {
-                    "query": search_term,
-                    "operator": "and"
-                }
-            }
-        }
-    else:
-        # QueryString Query
-        return {
-            "query_string": {
-                "query": search_term,
-                "fields": [
-                    field
-                ],
-                "default_operator": "AND"
-            }
-        }
-
-
 class BaseBuilder(object):
     __metaclass__ = abc.ABCMeta
 
@@ -233,19 +206,41 @@ class SearchBuilder(BaseBuilder):
         }
 
         # Highlight
-        if not self.params.get("no_highlight") and \
-                not self.params.get("size") == 0:
+        if not self.params.get("no_highlight"):
             search["highlight"] = self._build_highlight()
 
         # sort
-        if not self.params.get("size") == 0:
-            search["sort"] = self._build_sort()
+        search["sort"] = self._build_sort()
 
         # query
         search_term = self.params.get("search_term")
         if search_term:
-            search["query"] = build_search_terms(
-                search_term, self.params.get("field"))
+            if (re.match(r"^[A-Za-z\d\s]+$", search_term) and not
+                any(keyword in search_term
+                    for keyword in ("AND", "OR", "NOT", "TO"))):
+
+                # Match Query
+                search["query"] = {
+                    "match": {
+                        self.params.get("field"): {
+                            "query": search_term,
+                            "operator": "and"
+                        }
+                    }
+                }
+
+            else:
+
+                # QueryString Query
+                search["query"] = {
+                    "query_string": {
+                        "query": search_term,
+                        "fields": [
+                            self.params.get("field")
+                        ],
+                        "default_operator": "AND"
+                    }
+                }
 
         return search
 
@@ -427,6 +422,10 @@ class StateAggregationBuilder(BaseBuilder):
     def __init__(self):
         BaseBuilder.__init__(self)
         self.filter_clauses = None
+        self.exclude = []
+
+    def add_exclude(self, field_name_list):
+        self.exclude += field_name_list
 
     def build_one(self, field_name):
         # Lazy initialization
@@ -486,8 +485,16 @@ class StateAggregationBuilder(BaseBuilder):
         if company_filter:
             field_aggs["filter"]["bool"]["filter"].append(company_filter)
 
+        # Add filter clauses to aggregation entries (only those that are not
+        # the same as field name or part of the exclude list, which means we
+        # want to list all matched aggregation)
         for item in self.params:
-            if item in (
+            include_filter = (
+                item != field_name or
+                (item == field_name and item in self.exclude)
+            )
+
+            if include_filter and item in (
                 self._OPTIONAL_FILTERS + self._OPTIONAL_FILTERS_STRING_TO_BOOL
             ):
                 field_level_should = {
@@ -503,242 +510,12 @@ class StateAggregationBuilder(BaseBuilder):
         aggs = {}
 
         agg_fields = self._AGG_FIELDS
+        if self.exclude:
+            agg_fields = [field_name for field_name in self._AGG_FIELDS
+                          if field_name not in self.exclude]
         for field_name in agg_fields:
             aggs[field_name] = self.build_one(field_name)
 
-        return aggs
-
-
-class LensAggregationBuilder(BaseBuilder):
-    _ES_CHILD_AGG_MAP = {
-        'product.raw': 'sub-product',
-        'sub_product.raw': 'product',
-        'issue.raw': 'sub-issue',
-        'sub_issue.raw': 'issue',
-        'tags': 'tags'
-    }
-
-    def __init__(self):
-        super(LensAggregationBuilder, self).__init__()
-        self.filter_clauses = None
-
-    def build_histogram(self, agg_name, interval, include_date_filter):
-        agg = {
-            "aggs": {
-                agg_name: {
-                    "date_histogram": {
-                        "field": "date_received",
-                        "interval": interval
-                    }
-                }
-            }
-        }
-
-        # Lazy initialization
-        if not self.filter_clauses:
-            self.filter_clauses = self._build_filter_clauses()
-
-        filters = {
-            "filter": {
-                "bool": {
-                    "must": [],
-                    "should": [],
-                    "filter": [],
-                }
-            }
-        }
-
-        date_filter = self._build_date_range_filter(
-            self.params.get("date_received_min"),
-            self.params.get("date_received_max"), "date_received")
-
-        if date_filter and include_date_filter:
-            filters["filter"]["bool"]["filter"].append(date_filter)
-
-        company_filter = self._build_date_range_filter(
-            self.params.get("company_received_min"),
-            self.params.get("company_received_max"), "date_sent_to_company")
-
-        if company_filter:
-            filters["filter"]["bool"]["filter"].append(company_filter)
-
-        for item in self.params:
-            if item in (
-                self._OPTIONAL_FILTERS + self._OPTIONAL_FILTERS_STRING_TO_BOOL
-            ):
-                field_level_should = {
-                    "bool": {"should": self.filter_clauses[item]}
-                }
-                filters["filter"]["bool"]["filter"].append(
-                    field_level_should
-                )
-
-        agg['filter'] = filters['filter']
-
-        return agg
-
-    def date_extreme(self, extreme):
-        return {
-            extreme: {
-                "field": "date_received",
-                "format": "yyyy-MM-dd'T'12:00:00-05:00"
-            }
-        }
-
-
-class TrendsAggregationBuilder(LensAggregationBuilder):
-    _AGG_FIELDS = (
-        'tags',
-        'issue',
-        'product'
-    )
-
-    _AGG_HEADING_MAP = {
-        'tags': 'tags',
-        'issue': 'issue',
-        'product': 'product',
-        'sub_product': 'sub-product',
-        'sub_issue': 'sub-issue',
-    }
-
-    def __init__(self):
-        super(TrendsAggregationBuilder, self).__init__()
-
-    def percent_change_agg(self, es_field_name, interval, trend_depth):
-        return {
-            "terms": {
-                "field": es_field_name,
-                "size": trend_depth
-            },
-            "aggs": {
-                "trend_period": {
-                    "date_histogram": {
-                        "field": "date_received",
-                        "interval": interval
-                    },
-                    "aggs": {
-                        "interval_diff": {
-                            "serial_diff": {"buckets_path": "_count"}
-                        }
-                    }
-                }
-            }
-        }
-
-    def get_agg_heading(self, field_name):
-        return self._AGG_HEADING_MAP.get(field_name, field_name)
-
-    def agg_setup(self, field_name, agg_heading_name, interval):
-        field_aggs = {
-            "aggs": {}
-        }
-
-        es_field_name = self._OPTIONAL_FILTERS_PARAM_TO_ES_MAP.get(
-            field_name, field_name
-        )
-
-        field_aggs["aggs"][agg_heading_name] = self.percent_change_agg(
-            es_field_name, interval, self.params['trend_depth'])
-
-        # Lazy initialization
-        if not self.filter_clauses:
-            self.filter_clauses = self._build_filter_clauses()
-
-        filters = {
-            "filter": {
-                "bool": {
-                    "must": [],
-                    "should": [],
-                    "filter": [],
-                }
-            }
-        }
-
-        date_filter = self._build_date_range_filter(
-            self.params.get("date_received_min"),
-            self.params.get("date_received_max"), "date_received")
-
-        if date_filter:
-            filters["filter"]["bool"]["filter"].append(date_filter)
-
-        company_filter = self._build_date_range_filter(
-            self.params.get("company_received_min"),
-            self.params.get("company_received_max"), "date_sent_to_company")
-
-        if company_filter:
-            filters["filter"]["bool"]["filter"].append(company_filter)
-
-        for item in self.params:
-            if item in (
-                self._OPTIONAL_FILTERS + self._OPTIONAL_FILTERS_STRING_TO_BOOL
-            ):
-                field_level_should = {
-                    "bool": {"should": self.filter_clauses[item]}
-                }
-                filters["filter"]["bool"]["filter"].append(
-                    field_level_should
-                )
-
-        field_aggs['filter'] = filters['filter']
-
-        return field_aggs
-
-    def build_one_overview(self, field_name, agg_heading_name, interval):
-        field_aggs = self.agg_setup(field_name, agg_heading_name, interval)
-
-        if field_name in self._OPTIONAL_FILTERS_CHILD_MAP:
-            es_child_name = self._OPTIONAL_FILTERS_PARAM_TO_ES_MAP.get(
-                self._OPTIONAL_FILTERS_CHILD_MAP.get(field_name))
-            child_agg_name = self._ES_CHILD_AGG_MAP.get(es_child_name)
-            field_aggs["aggs"][agg_heading_name]["aggs"][child_agg_name] = \
-                self.percent_change_agg(es_child_name, interval, 0)
-
-        return field_aggs
-
-    def build_one_lens(self, field_name, agg_heading_name, interval, sub_lens):
-        field_aggs = self.agg_setup(field_name, agg_heading_name, interval)
-
-        es_child_name = self._OPTIONAL_FILTERS_PARAM_TO_ES_MAP.get(
-            sub_lens)
-        child_agg_name = self._ES_CHILD_AGG_MAP.get(es_child_name)
-        field_aggs["aggs"][agg_heading_name]["aggs"][child_agg_name] = \
-            self.percent_change_agg(es_child_name, interval,
-                                    self.params['sub_lens_depth'])
-
-        return field_aggs
-
-    def build(self):
-        aggs = {}
-
-        aggs['dateRangeArea'] = self.build_histogram(
-            'dateRangeArea', self.params['trend_interval'], True)
-        aggs['dateRangeBrush'] = self.build_histogram(
-            'dateRangeBrush', self.params['trend_interval'], False)
-
-        if self.params['lens'] == 'overview':
-            # Reset default for overview row charts
-            self.params['trend_depth'] = 5
-
-            for field_name in self._AGG_FIELDS:
-                agg_heading_name = self.get_agg_heading(field_name)
-
-                aggs[agg_heading_name] = self.build_one_overview(
-                    field_name,
-                    agg_heading_name,
-                    self.params['trend_interval']
-                )
-        else:
-            agg_heading_name = self.get_agg_heading(self.params['lens'])
-
-            aggs[agg_heading_name] = self.build_one_lens(
-                self.params['lens'],
-                agg_heading_name,
-                self.params['trend_interval'],
-                self.params['sub_lens']
-            )
-
-        aggs['min_date'] = self.date_extreme('min')
-        aggs['max_date'] = self.date_extreme('max')
         return aggs
 
 
@@ -751,5 +528,3 @@ if __name__ == "__main__":
     print(aggbuilder.build())
     stateaggbuilder = StateAggregationBuilder()
     print(stateaggbuilder.build())
-    trendsaggbuilder = TrendsAggregationBuilder()
-    print(trendsaggbuilder.build())
