+++ complaint_search/es_interface.py
@@ -13,7 +13,6 @@ from complaint_search.es_builders import (
     PostFilterBuilder,
     SearchBuilder,
     StateAggregationBuilder,
-    TrendsAggregationBuilder,
 )
 from complaint_search.export import ElasticSearchExporter
 from elasticsearch import Elasticsearch, helpers
@@ -29,81 +28,6 @@ _ES_INSTANCE = None
 
 _COMPLAINT_ES_INDEX = os.environ.get('COMPLAINT_ES_INDEX', 'complaint-index')
 _COMPLAINT_DOC_TYPE = os.environ.get('COMPLAINT_DOC_TYPE', 'complaint-doctype')
-
-
-# -----------------------------------------------------------------------------
-# Trends Operations
-# -----------------------------------------------------------------------------
-
-# Safely extracts a string from an object
-# if it isn't present, the default value is returned
-def extract_date(agg_term, default_value):
-    return agg_term['value_as_string'] \
-        if 'value_as_string' in agg_term else default_value
-
-
-def build_trend_meta(response):
-    meta = {}
-
-    if 'max_date' in response['aggregations']:
-        date_max = extract_date(response['aggregations']['max_date'], None)
-        date_min = extract_date(response['aggregations']['min_date'], None)
-
-        # del response["aggregations"]["max_date"]
-        # del response["aggregations"]["min_date"]
-
-        meta['date_min'] = date_min
-        meta['date_max'] = date_max
-
-    return meta
-
-
-# Find sub_agg name if it exists in order to filter percent change
-def get_sug_agg_key_if_exists(agg):
-    agg_keys_exclude = ('trend_period', 'key', 'doc_count')
-    for key in agg.keys():
-        if key not in agg_keys_exclude:
-            return key
-    return None
-
-
-# Filter out all but the most recent buckets in sub agg for the Percent
-# Change on chart
-def process_trend_aggregations(aggregations):
-    trend_charts = (
-        'product',
-        'sub-product',
-        'issue',
-        'sub-issue',
-        'tags'
-    )
-
-    for agg_name in trend_charts:
-        if agg_name in aggregations:
-            agg_buckets = \
-                aggregations[agg_name][agg_name]['buckets']
-            for sub_agg in agg_buckets:
-                sub_agg['trend_period']['buckets'] = sorted(
-                    sub_agg['trend_period']['buckets'],
-                    key=lambda k: k['key_as_string'], reverse=True)
-                sub_agg_name = get_sug_agg_key_if_exists(sub_agg)
-                if sub_agg_name:
-                    for sub_sub_agg in sub_agg[sub_agg_name]['buckets']:
-                        sub_sub_agg['trend_period']['buckets'] = sorted(
-                            sub_sub_agg['trend_period']['buckets'],
-                            key=lambda k: k['key_as_string'],
-                            reverse=True)[1:2]
-    return aggregations
-
-
-# Process the response from a trends query
-def process_trends_response(response):
-    response['aggregations'] = \
-        process_trend_aggregations(response['aggregations'])
-
-    response['_meta'] = build_trend_meta(response)
-
-    return response
 
 
 def _get_es():
@@ -376,7 +300,7 @@ def document(complaint_id):
     return res
 
 
-def states_agg(**kwargs):
+def states_agg(agg_exclude=None, **kwargs):
     params = copy.deepcopy(PARAMS)
     params.update(**kwargs)
     search_builder = SearchBuilder()
@@ -389,37 +313,21 @@ def states_agg(**kwargs):
         _ES_URL, _COMPLAINT_ES_INDEX, _COMPLAINT_DOC_TYPE, body
     )
 
+    # AD, TODO: Do I need to hardcode?
     body["size"] = 0
+    del body['_source']
+    del body['highlight']
+    del body['sort']
+
     aggregation_builder = StateAggregationBuilder()
     aggregation_builder.add(**params)
+    if agg_exclude:
+        aggregation_builder.add_exclude(agg_exclude)
     body["aggs"] = aggregation_builder.build()
 
     res = _get_es().search(index=_COMPLAINT_ES_INDEX,
                            doc_type=_COMPLAINT_DOC_TYPE,
                            body=body,
                            scroll="10m")
-
-    return res
-
-
-def trends(**kwargs):
-    params = copy.deepcopy(PARAMS)
-    params.update(**kwargs)
-    params.update(size=0)
-    search_builder = SearchBuilder()
-    search_builder.add(**params)
-    body = search_builder.build()
-
-    res = None
-
-    aggregation_builder = TrendsAggregationBuilder()
-    aggregation_builder.add(**params)
-    body["aggs"] = aggregation_builder.build()
-
-    res = _get_es().search(index=_COMPLAINT_ES_INDEX,
-                           doc_type=_COMPLAINT_DOC_TYPE,
-                           body=body)
-
-    res = process_trends_response(res)
 
     return res
