+++ swagger-config.yaml
@@ -40,7 +40,7 @@ paths:
         - $ref: '#/components/parameters/product'
         - $ref: '#/components/parameters/state'
         - $ref: '#/components/parameters/submitted_via'
-        - $ref: '#/components/parameters/tags'
+        - $ref: '#/components/parameters/tag'
         - $ref: '#/components/parameters/timely'
         - $ref: '#/components/parameters/zip_code'
       responses:
@@ -95,7 +95,7 @@ paths:
         - $ref: '#/components/parameters/product'
         - $ref: '#/components/parameters/state'
         - $ref: '#/components/parameters/submitted_via'
-        - $ref: '#/components/parameters/tags'
+        - $ref: '#/components/parameters/tag'
         - $ref: '#/components/parameters/timely'
         - $ref: '#/components/parameters/zip_code'
       responses:
@@ -129,7 +129,7 @@ paths:
         - $ref: '#/components/parameters/product'
         - $ref: '#/components/parameters/state'
         - $ref: '#/components/parameters/submitted_via'
-        - $ref: '#/components/parameters/tags'
+        - $ref: '#/components/parameters/tag'
         - $ref: '#/components/parameters/timely'
         - $ref: '#/components/parameters/zip_code'
       responses:
@@ -191,7 +191,7 @@ paths:
         - $ref: '#/components/parameters/product'
         - $ref: '#/components/parameters/state'
         - $ref: '#/components/parameters/submitted_via'
-        - $ref: '#/components/parameters/tags'
+        - $ref: '#/components/parameters/tag'
         - $ref: '#/components/parameters/timely'
         - $ref: '#/components/parameters/zip_code'
       responses:
@@ -201,54 +201,9 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/StatesResult'
-  /trends:
-    get:
-      tags:
-        - Trends
-      summary: "List trends"
-      description: "Return specific aggregations for a search"
-      parameters:
-        - $ref: '#/components/parameters/search_term'
-        - $ref: '#/components/parameters/field'
-        - $ref: '#/components/parameters/company'
-        - $ref: '#/components/parameters/company_public_response'
-        - $ref: '#/components/parameters/company_received_max'
-        - $ref: '#/components/parameters/company_received_min'
-        - $ref: '#/components/parameters/company_response'
-        - $ref: '#/components/parameters/consumer_consent_provided'
-        - $ref: '#/components/parameters/consumer_disputed'
-        - $ref: '#/components/parameters/date_received_max'
-        - $ref: '#/components/parameters/date_received_min'
-        - $ref: '#/components/parameters/has_narrative'
-        - $ref: '#/components/parameters/issue'
-        - $ref: '#/components/parameters/lens'
-        - $ref: '#/components/parameters/product'
-        - $ref: '#/components/parameters/state'
-        - $ref: '#/components/parameters/submitted_via'
-        - $ref: '#/components/parameters/sub_lens'
-        - $ref: '#/components/parameters/sub_lens_depth'
-        - $ref: '#/components/parameters/tags'
-        - $ref: '#/components/parameters/timely'
-        - $ref: '#/components/parameters/trend_depth'
-        - $ref: '#/components/parameters/trend_interval'
-        - $ref: '#/components/parameters/zip_code'
-      responses:
-        '200':
-          description: successful operation
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/TrendsResult'
-            text/csv:
-              schema:
-                $ref: '#/components/schemas/TrendsResult'
-        '400':
-          description: Invalid status value
 tags:
   - name: Complaints
     description: These endpoints provide access to consumer complaints
-  - name: Trends
-    description: These endpoints provide access aggregated consumer complaint data
   - name: Typeahead
     description: These endpoints support the typeahead boxes in the UI
 externalDocs:
@@ -380,22 +335,6 @@ components:
         type: array
         items:
           type: string
-    lens:
-      name: lens
-      in: query
-      required: true
-      description: The data lens through which to view complaint trends over time.
-      explode: true
-      schema:
-        type: string
-        items:
-          type: string
-        enum:
-          - overview
-          - issue
-          - product
-          - tags
-        default: overview
     no_aggs:
       name: no_aggs
       in: query
@@ -456,31 +395,6 @@ components:
         type: array
         items:
           type: string
-    sub_lens:
-      name: sub_lens
-      in: query
-      description: The sub-lens through which to view complaint trends over time.
-      explode: true
-      schema:
-        type: string
-        items:
-          type: string
-        enum:
-          - issue
-          - product
-          - sub_product
-          - sub_issue
-          - tags
-    sub_lens_depth:
-      name: sub_lens_depth
-      in: query
-      description: The top X trend sub aggregations will be returned, where X is the supplied sub_lens_depth.
-      schema:
-        type: integer
-        maximum: 10000000
-        minimum: 5
-        format: int64
-        default: 10
     submitted_via:
       name: submitted_via
       in: query
@@ -497,7 +411,7 @@ components:
       required: true
       schema:
         type: string
-    tags:
+    tag:
       name: tags
       in: query
       description: Filter the results to only return these types of tag
@@ -515,29 +429,6 @@ components:
         type: array
         items:
           type: string
-    trend_depth:
-      name: trend_depth
-      in: query
-      description: The top X trend aggregations will be returned, where X is the supplied trend_depth.
-      schema:
-        type: integer
-        maximum: 10000000
-        minimum: 5
-        format: int64
-        default: 10
-    trend_interval:
-      name: trend_interval
-      in: query
-      description: The interval of time to use for trends aggregations histograms.
-      required: true
-      schema:
-        type: string
-        enum:
-          - yearly
-          - quarterly
-          - monthly
-          - weekly
-          - daily
     zip_code:
       name: zip_code
       in: query
@@ -783,22 +674,6 @@ components:
       type: array
       items:
         type: string
-    TrendsResult:
-      type: object
-      properties:
-        aggregations:
-          type: object
-          properties:
-            issue:
-              $ref: '#/components/schemas/MultiLevelAggregation'
-            product:
-              $ref: '#/components/schemas/MultiLevelAggregation'
-            sub_issue:
-              $ref: '#/components/schemas/MultiLevelAggregation'
-            sub_product:
-              $ref: '#/components/schemas/MultiLevelAggregation'
-            tags:
-              $ref: '#/components/schemas/MultiLevelAggregation'
   links: {}
   callbacks: {}
 security: []
